bp: Concurrent file chunk fetching for CCR restore
https://github.com/elastic/elasticsearch/commit/9b75a709a2c3734187cc49762b9ef5d1b18f58dd

This doesn't include the CCR part, which we don't have and we couldn't
copy because of the different license.
